Change to npm as package manager for now
Using yarn, specially the newer (4.X) is very hacky in the hermetic env
we need to build the image. Since there is no base image with yarn in
it, we would need to install yarn first in some prefetch step, and make
it available during the build.
We prefetch all the dependencies now with npm, and make those available
with cachi2 (prefetch task)