Keep operatorframework.io/initialization-resource in sync
We add a single yq/sed command to make sure we sync the content from
config/samples/scale_v1alpha1_storagescale.yaml into the
operatorframework.io/initialization-resource.

We do this because otherwise a user will get different values depending
on where he clicks in the UI when creating the StorageScale CR.
(The normal UI for it vs the "required Storage button")

We add it to the manifests target as any of the early targets will do.

We use sed for the replacement because using yq triggers a bunch of
forced reindentation which we do not want at this point: yq forces it
to be two spaces everywhere, but the bundle reverts that in the bundles.
So I prefer to keep this a bit smaller/simpler at this time.