List all the devices we know of by default
Starting with v5.2.3.x symlinks are finally supported this made us add
/dev/disk/by-id/* in our default cluster config [1]

The problem with this is that once you specify an entry in `localDevicePaths`
no other device will be accepted. So by default let's list all the
currently documented accepted devices in addition to /dev/disk/by-id/*.

Tested and this fixed the partially-shared issues we were observing.

[1] https://www.ibm.com/docs/en/scalecontainernative/5.2.2?topic=systems-local-file-system#specify-device-names