Drop MachineConfig creation support
There is little gain from this operator in adding machineconfig for the
kernel-devel. The hope is still that this is not needed in the future.
It might be that for the first release we will ask the user to create
this MachineConfig and then later we land KMM support fully.