Drop the pull secret creation
At this stage we expect the user to patch the global pull secret.
We might amend this in the future (especially once KMM releases version
2.4 and works properly with secrets)

For now I am not dropping the pullsecret references in the build
scaffolding (Makefile, hack/build.sh) in case we quickly need
something in the short term.