Fix udev triggers in devicefinder
The udevadm is completetly missing from the used image, and the usage of
stdbuf in exec.Run made the command fail with silently. Udev chanages
were not monitored, and not triggering a rescan of devices due to this
silent error.
Removed the usage of wrapped stdbuf, since we already using bufio in the
golang code to parse the output.
Added the missing udev package to the docker file.
NOTE: this will break if we move to build and release with konflux, but
releasing from locally will work for now, and unblock some UI issues.